chat class update
