+++ static/js/blamer.js
@@ -142,31 +142,6 @@ export default class extends abstract {
         return container;
 
     }
-    unseenMessage(message) {
-        if(message.receiver == this.data.NickName && message.seen === 0){
-            return true
-        }
-        return false
-    }
-    findChatBox(receiver) {
-        const regex = /(chatWith)_(\w+)/;
-
-        const match = receiver.match(regex);
-        // const chatWith = match[1]; // "chatWith"
-        const id = match[2]; // "testUser7"
-
-        let container = document.createElement("div");
-        container.id = "bChatBox";
-        container.className = "bChatBox";
-        container.innerHTML = `
-              <div class="cReceiverName"><span id="receiverName">${id}</span></div>
-              <div id="message-list"class="cArea"> </div>
-              <div class="cInput">
-              <input type="text" id="message-input" placeholder="Type your message here">
-              </div>
-              `
-        return container;
-    }
     updatedPostList(topic) {
 
         let postBox = document.querySelectorAll(".bPost")
@@ -282,119 +257,7 @@ export default class extends abstract {
         `;
         return postBox;
     }
-    async updatedChat(receiver) {
-        console.log("im chating with" + receiver)
-        let chat = "";
-        let messages = [];
-        if (this.message) {
-            if (this.message.sender == this.data.NickName)
-                if (this.data.Messages.send){
-                    this.data.Messages.send =
-                    this.data.Messages.send.concat(this.message);
-                }else {
-                    this.data.Messages.send = [this.message];
-                }
-            else if (this.message.receiver == this.data.NickName)
-                if (this.data.Messages.receive){
-                    this.data.Messages.receive =
-                    this.data.Messages.receive.concat(
-                        this.message
-                );
-                }else {
-                    this.data.Messages.receive = [this.message];
-                }
-        }
-        // Combine send and receive messages into a single array
-        if (this.data.Messages.send) {
-            messages = messages.concat(this.data.Messages.send);
-        }
-        if (this.data.Messages.receive) {
-            messages = messages.concat(this.data.Messages.receive);
-        }
-
-        // Sort messages by time
-        messages.sort((a, b) => new Date(a.time) - new Date(b.time));
-        // Generate chat HTML
-
-        messages.forEach((message) => {
-            if (receiver == null) {
-                if (((message.sender == this.data.NickName && message.receiver == receiver) ||
-                        (message.receiver == this.data.NickName && message.sender == receiver)) &&
-                    message.content.length > 39
-                ) {
-                    for (
-                        let i = 0; i < message.content.length; i += 40
-                    ) {
-                        message.content =
-                            message.content.slice(0, i) +
-                            "\n" +
-                            message.content.slice(i);
-                    }
-                }
-            }
-            if (
-                message.sender == this.data.NickName &&
-                message.receiver == receiver
-            ) {
-                chat += `
-                            <div class="messageBox" style="justify-items: end;">
-                            <div class="mInfo" style="float:right;"><b>Me:</b> ${message.time}</div>
-                            <div class=" message"><span>${message.content}</span></div>
-                            </div>
-                  `;
-            } else if (
-                message.receiver == this.data.NickName &&
-                message.sender == receiver
-            ) {
-                chat += `
-                            <div class="messageBox" style="justify-items: start;">
-                            <div class="mInfo" style="float:left; "><b>${message.sender}:</b> ${message.time}</div>
-                            <div class=" message"><span>${message.content}</span></div>
-                            </div>`;
-            }
-        });
-        return chat;
-    }
-    // updatedChatBox return updated chatbox
-    async updatedChatBox(message) {
-        console.log("im chating with message is ???",message)
-        if (
-            this.activeUserName != null &&
-            this.activeUserName !== "guest"
-        ) {
-            if (message != undefined) {
-
-                this.message = message;
-                let parent =
-                    document.getElementById("message-list");
-                let children =
-                    document.querySelectorAll(".message");
-                if (children.length > 0) {
-                    children.forEach((child) => child.remove());
-                }
-                if (
-                    this.message.sender ==
-                    this.activeUserName ||
-                    this.message.receiver ==
-                    this.activeUserName
-                ) {
-                    parent.innerHTML = await this.updatedChat(document.getElementById("receiverName").textContent);
-                }
-            } else {
-                let parent =
-                    document.getElementById("message-list");
-                let children =
-                    document.querySelectorAll(".message");
-                if (children.length > 0) {
-                    children.forEach((child) => child.remove());
-                }
-                if (
-                    this.data.NickName ===
-                    this.activeUserName
-                ) {
-                    parent.innerHTML = await this.updatedChat(document.getElementById("receiverName").textContent);
-                }
-            }
-        }
+    getMessages(){
+        return this.data.Messages;
     }
 }
